+++ src/cosmosdb-preview/azext_cosmosdb_preview/_client_factory.py
@@ -115,3 +115,7 @@ def cf_fleetspace(cli_ctx, _):
 
 def cf_fleetspace_account(cli_ctx, _):
     return cf_cosmosdb_preview(cli_ctx).fleetspace_account
+
+
+def cf_fleet_analytics(cli_ctx, _):
+    return cf_cosmosdb_preview(cli_ctx).fleet_analytics